refactor(client): extract config/db/health/learning into internal packages
Move four subsystems from monolithic package main into typed internal packages:
- internal/config  — Config, ServerConfig, CloudConfig, env accessors, loadenv
- internal/db      — Database, all record types, migrations
- internal/health  — HealthSnapshot and process-health helpers
- internal/learning — LearningConsent, license manager, EnsureTermsAccepted

Ten original files emptied to `package main` stubs; shim files
(config_shim.go, db_shim.go, health_shim.go, learning_shim.go) expose
type aliases and function forwards so zero call-sites change.

CLI handlers split out:
- cloud_cli.go   — `devtrack cloud login/logout/status` (was cloud.go)
- license_cli.go — `devtrack login/logout/license/terms/telemetry` (was license.go)

main.go: inject ldflags Version into internal/config via cfg.SetBuildVersion.

Build verified: `go build ./...` and `go vet ./...` both exit 0.
